requirements.txt
django
djangorestframework
selenium
pandas
gunicorn
webdriver_manager

